+++ src/types/index.d.ts
@@ -1,3 +1,4 @@
+export type GetTasksAPIParamTypes = "completed" | "uncompleted" | "all";
 export type Task = {
   id: number;
   /**
@@ -18,12 +19,6 @@ export type Task = {
    */
   updated_at: string;
 };
-export type TaskResBody<
-  Data = unknown,
-  HasTotal extends boolean = false,
-> = HasTotal extends boolean
-  ? (TaskResBodySuccess<Data> & { total: number }) | TaskResBodyError
-  : TaskResBodySuccess<Data> | TaskResBodyError;
 export type TaskResBodySuccess<Data = unknown> = {
   status: "success";
   data: Data;
@@ -32,12 +27,14 @@ export type TaskResBodyError = {
   status: "error";
   message: string;
 };
-export type GetTasksResBody = TaskResBody<Task[], true>;
+export type GetTasksResBody =
+  | (TaskResBodySuccess<Task[]> & { total: number })
+  | TaskResBodyError;
 export type PostTaskReqBody = Pick<
   Task,
   "name" | "description" | "is_completed" | "created_at" | "updated_at"
 >;
-export type PostTaskResBody = TaskResBody<Task>;
+export type PostTaskResBody = TaskResBodySuccess<Task> | TaskResBodyError;
 export type PutTaskReqBody = Pick<Task, "name" | "description" | "updated_at">;
-export type PutTaskResBody = TaskResBody<Task>;
-export type PatchTaskResBody = TaskResBody<Task>;
+export type PutTaskResBody = TaskResBodySuccess<Task> | TaskResBodyError;
+export type PatchTaskResBody = TaskResBodySuccess<Task> | TaskResBodyError;